Implement `which-formula`
