+++ Library/Homebrew/sorbet/rbi/dsl/homebrew/cmd/which_formula.rbi
@@ -1,0 +1,19 @@
+# typed: true
+
+# DO NOT EDIT MANUALLY
+# This is an autogenerated file for dynamic methods in `Homebrew::Cmd::WhichFormula`.
+# Please instead update this file by running `bin/tapioca dsl Homebrew::Cmd::WhichFormula`.
+
+
+class Homebrew::Cmd::WhichFormula
+  sig { returns(Homebrew::Cmd::WhichFormula::Args) }
+  def args; end
+end
+
+class Homebrew::Cmd::WhichFormula::Args < Homebrew::CLI::Args
+  sig { returns(T::Boolean) }
+  def explain?; end
+
+  sig { returns(T::Boolean) }
+  def skip_update?; end
+end
